+++ PgRoutiner/DataAccess/GetRoutineCount.cs
@@ -6,15 +6,12 @@ namespace PgRoutiner.DataAccess;
 public static partial class DataAccessConnectionExtensions
 {
     public static long GetRoutineCount(this NpgsqlConnection connection, Settings settings, 
-        string schemaSimilarTo = null, string schemaNotSimilarTo = null) =>
-        connection
-        .WithParameters(new
-        {
-            schema = (schemaSimilarTo ?? settings.SchemaSimilarTo, DbType.AnsiString),
-            not_schema = (schemaNotSimilarTo ?? settings.SchemaNotSimilarTo, DbType.AnsiString),
-            notSimilarTo = (settings.RoutinesNotSimilarTo, DbType.AnsiString),
-            similarTo = (settings.RoutinesSimilarTo, DbType.AnsiString)
-        })
+        string schemaSimilarTo = null, string schemaNotSimilarTo = null) => connection
+        .WithParameters(
+            (schemaSimilarTo ?? settings.SchemaSimilarTo, DbType.AnsiString),
+            (schemaNotSimilarTo ?? settings.SchemaNotSimilarTo, DbType.AnsiString),
+            (settings.RoutinesNotSimilarTo, DbType.AnsiString),
+            (settings.RoutinesSimilarTo, DbType.AnsiString))
         .Read<long>(@$"
 
                 select 
@@ -24,12 +21,12 @@ public static partial class DataAccessConnectionExtensions
                 where
                     lower(r.external_language) = any('{{sql,plpgsql}}')
                     and
-                    (   @schema is null or (r.specific_schema similar to @schema)   )
-                    and (   @not_schema is null or r.specific_schema not similar to @not_schema   )
+                    (   $1 is null or (r.specific_schema similar to $1)   )
+                    and (   $2 is null or r.specific_schema not similar to $2   )
                     and (   {GetSchemaExpression("r.specific_schema")}  )
 
-                    and (@notSimilarTo is null or r.routine_name not similar to @notSimilarTo)
-                    and (@similarTo is null or r.routine_name similar to @similarTo)
+                    and ($3 is null or r.routine_name not similar to $3)
+                    and ($4 is null or r.routine_name similar to $4)
             ")
             .Single();
 }