+++ PgRoutiner/DataAccess/GetRoutines.cs
@@ -9,11 +9,9 @@ public static partial class DataAccessConnectionExtensions
     public static IEnumerable<PgItem> GetRoutines(this NpgsqlConnection connection, Settings settings)
     {
         return connection
-            .WithParameters(new
-            {
-                schema = (settings.SchemaSimilarTo, DbType.AnsiString),
-                not_schema = (settings.SchemaNotSimilarTo, DbType.AnsiString)
-            })
+            .WithParameters(
+                (settings.SchemaSimilarTo, DbType.AnsiString),
+                (settings.SchemaNotSimilarTo, DbType.AnsiString))
             .Read<(string Schema, string Name, string Type)>(@$"
 
                 select 
@@ -26,8 +24,8 @@ public static partial class DataAccessConnectionExtensions
                 where
                     lower(r.external_language) = any('{{sql,plpgsql}}')
                     and
-                    (   @schema is null or (r.specific_schema similar to @schema)   )
-                    and (   @not_schema is null or r.specific_schema not similar to @not_schema   )
+                    (   $1 is null or (r.specific_schema similar to $1)   )
+                    and (   $2 is null or r.specific_schema not similar to $2   )
                     and (   {GetSchemaExpression("r.specific_schema")}  )