+++ PgRoutiner/DataAccess/GetTables.cs
@@ -9,12 +9,10 @@ public static partial class DataAccessConnectionExtensions
     public static IEnumerable<PgItem> GetTables(this NpgsqlConnection connection, Settings settings, string skipSimilar = null)
     {
         return connection
-            .WithParameters(new
-            {
-                schema = (settings.SchemaSimilarTo, DbType.AnsiString),
-                not_schema = (settings.SchemaNotSimilarTo, DbType.AnsiString),
-                skipSimilar = (skipSimilar, DbType.AnsiString),
-            })
+            .WithParameters(
+                (settings.SchemaSimilarTo, DbType.AnsiString),
+                (settings.SchemaNotSimilarTo, DbType.AnsiString),
+                (skipSimilar, DbType.AnsiString))
             .Read<(string Schema, string Name, string Type)>(@$"
 
         select 
@@ -22,10 +20,10 @@ public static partial class DataAccessConnectionExtensions
         from 
             information_schema.tables
         where
-            (   @schema is null or (table_schema similar to @schema)   )
-            and (   @not_schema is null or (table_schema not similar to @not_schema)   )
+            (   $1 is null or (table_schema similar to $1)   )
+            and (   $2 is null or (table_schema not similar to $2)   )
             and (   {GetSchemaExpression("table_schema")}  )
-            and (   @skipSimilar is null or (table_name not similar to @skipSimilar)   )
+            and (   $3 is null or (table_name not similar to $3)   )
 
         ")
         .Select(t => new PgItem